+++ ta4j-core/src/main/java/org/ta4j/core/BacktestExecutor.java
@@ -24,8 +24,8 @@
 package org.ta4j.core;
 
 import org.ta4j.core.num.Num;
-import org.ta4j.core.report.TradingStatement;
-import org.ta4j.core.report.TradingStatementGenerator;
+import org.ta4j.core.reports.TradingStatement;
+import org.ta4j.core.reports.TradingStatementGenerator;
 
 import java.util.ArrayList;
 import java.util.List;
