+++ ta4j-core/src/main/java/org/ta4j/core/reports/PerformanceReport.java
@@ -21,35 +21,41 @@
  * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package org.ta4j.core.report;
+package org.ta4j.core.reports;
 
 import org.ta4j.core.num.Num;
 
 /**
- * This class represents report with statistics for executed trades
+ * This class represents report which contains performance statistics
  */
-public class TradeStatsReport {
+public class PerformanceReport {
 
-    private final Num profitTradeCount;
-    private final Num lossTradeCount;
-    private final Num breakEvenTradeCount;
+    private final Num totalProfitLoss;
+    private final Num totalProfitLossPercentage;
+    private final Num totalProfit;
+    private final Num totalLoss;
 
-    public TradeStatsReport(Num profitTradeCount, Num lossTradeCount, Num breakEvenTradeCount) {
-        this.profitTradeCount = profitTradeCount;
-        this.lossTradeCount = lossTradeCount;
-        this.breakEvenTradeCount = breakEvenTradeCount;
+    public PerformanceReport(Num totalProfitLoss, Num totalProfitLossPercentage, Num totalProfit, Num totalLoss) {
+        this.totalProfitLoss = totalProfitLoss;
+        this.totalProfitLossPercentage = totalProfitLossPercentage;
+        this.totalProfit = totalProfit;
+        this.totalLoss = totalLoss;
     }
 
-    public Num getProfitTradeCount() {
-        return profitTradeCount;
+    public Num getTotalProfitLoss() {
+        return totalProfitLoss;
     }
 
-    public Num getLossTradeCount() {
-        return lossTradeCount;
+    public Num getTotalProfitLossPercentage() {
+        return totalProfitLossPercentage;
     }
 
-    public Num getBreakEvenTradeCount() {
-        return breakEvenTradeCount;
+    public Num getTotalProfit() {
+        return totalProfit;
+    }
+
+    public Num getTotalLoss() {
+        return totalLoss;
     }
 
 }
